Typecast COUNT_OF to int to ignore unsigned comparison warnings

+++ vest/constants.hpp
@@ -1,7 +1,8 @@
 #ifndef CONSTANTS_HPP
 #define CONSTANTS_HPP
 
-#define COUNT_OF(x) (sizeof((x)) / sizeof(0[(x)]))
+// Cast to int so that I don't get unsigned comparison warnings
+#define COUNT_OF(x) static_cast<int>(sizeof((x)) / sizeof(0[(x)]))
 
 // Generated from offsets.py
 #include "offsets.hpp"
